Fixed export sheet logic for vscode, removed unused file
